Test fixed, Package fixed, new test added
