--- api/secrets-manager-api/src/main/java/api/GenericManager.java
@@ -1,8 +1,0 @@
-package api;
-
-import exceptions.SecretException;
-
-public interface GenericManager {
-    String getSecret(String secretName) throws SecretException;
-    <T> T getSecret(String secretName, Class<T> cls) throws SecretException;
-}
