--- api/secrets-manager-api/src/main/java/api/GenericManagerAsync.java
@@ -1,9 +1,0 @@
-package api;
-
-import exceptions.SecretException;
-import reactor.core.publisher.Mono;
-
-public interface GenericManagerAsync {
-    Mono<String> getSecret(String secretName) throws SecretException;
-    <T> Mono<T> getSecret(String secretName, Class<T> cls) throws SecretException;
-}
