+++ api/secrets-manager-api/src/main/java/co/com/bancolombia/secretsmanager/api/GenericManager.java
@@ -1,0 +1,8 @@
+package co.com.bancolombia.secretsmanager.api;
+
+import co.com.bancolombia.secretsmanager.api.exceptions.SecretException;
+
+public interface GenericManager {
+    String getSecret(String secretName) throws SecretException;
+    <T> T getSecret(String secretName, Class<T> cls) throws SecretException;
+}
