+++ api/secrets-manager-api/src/main/java/co/com/bancolombia/secretsmanager/api/GenericManagerAsync.java
@@ -1,0 +1,9 @@
+package co.com.bancolombia.secretsmanager.api;;
+
+import co.com.bancolombia.secretsmanager.api.exceptions.SecretException;
+import reactor.core.publisher.Mono;
+
+public interface GenericManagerAsync {
+    Mono<String> getSecret(String secretName) throws SecretException;
+    <T> Mono<T> getSecret(String secretName, Class<T> cls) throws SecretException;
+}
