+++ api/secrets-manager-api/src/main/java/co/com/bancolombia/secretsmanager/api/exceptions/SecretException.java
@@ -1,0 +1,15 @@
+package co.com.bancolombia.secretsmanager.api.exceptions;
+
+/**
+ * Represents an Secret Exception.
+ *
+ * @author <a href="mailto:andmagom@bancolombia.com.co">Andrés Mauricio Gómez
+ * P.</a>
+ */
+public class SecretException extends Exception {
+
+    public SecretException(String message) {
+        super(message);
+    }
+
+}
