--- api/secrets-manager-api/src/main/java/exceptions/SecretException.java
@@ -1,15 +1,0 @@
-package exceptions;
-
-/**
- * Represents an Secret Exception.
- *
- * @author <a href="mailto:andmagom@bancolombia.com.co">Andrés Mauricio Gómez
- * P.</a>
- */
-public class SecretException extends Exception {
-
-    public SecretException(String message) {
-        super(message);
-    }
-
-}
