+++ async/aws-secrets-manager-async/aws-secrets-manager-async.gradle
@@ -16,7 +16,6 @@ dependencies {
     implementation 'software.amazon.awssdk:sts'
     implementation group: 'io.projectreactor', name: 'reactor-core', version: '3.3.10.RELEASE'
     compile group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.13.3'
-    testCompile group: 'junit', name: 'junit', version: '4.12'
 }
