+++ async/aws-secrets-manager-async/src/main/java/co/com/bancolombia/secretsmanager/model/AWSSecretDBModel.java
@@ -1,0 +1,20 @@
+package co.com.bancolombia.secretsmanager.model;
+
+import lombok.Data;
+import co.com.bancolombia.secretsmanager.utils.GsonUtils;
+
+@Data
+public final class AWSSecretDBModel {
+
+	private String username;
+	private String password;
+	private String host;
+	private String port;
+	private String dbname;
+	private String engine;
+	private String dbInstanceIdentifier;
+
+	public static AWSSecretDBModel getModel(String data) {
+		return GsonUtils.getInstance().stringToModel(data, AWSSecretDBModel.class);
+	}
+}
