+++ async/aws-secrets-manager-async/src/main/java/co/com/bancolombia/secretsmanager/utils/GsonUtils.java
@@ -1,0 +1,39 @@
+package co.com.bancolombia.secretsmanager.utils;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+
+/**
+ * Singleton - Gson
+ *
+ * @author jhagutie@bancolombia.com.co IAS
+ * @version 1.0.0
+ * @since JDK 1.8
+ */
+public class GsonUtils {
+
+    private final Gson gson;
+
+    private GsonUtils() {
+        gson = new GsonBuilder()
+                .create();
+    }
+
+    public static GsonUtils getInstance() {
+        return GsonUtilsHolder.INSTANCE;
+    }
+
+    private static class GsonUtilsHolder {
+        private static final GsonUtils INSTANCE = new GsonUtils();
+    }
+
+    /**
+     * @param data  JSON string
+     * @param model Entity to convert
+     * @param <T>   object
+     * @return T object
+     */
+    public <T> T stringToModel(String data, Class<T> model) {
+        return gson.fromJson(data, model);
+    }
+}
