--- async/aws-secrets-manager-async/src/main/java/model/AWSSecretDBModel.java
@@ -1,20 +1,0 @@
-package model;
-
-import lombok.Data;
-import utils.GsonUtils;
-
-@Data
-public final class AWSSecretDBModel {
-
-	private String username;
-	private String password;
-	private String host;
-	private String port;
-	private String dbname;
-	private String engine;
-	private String dbInstanceIdentifier;
-
-	public static AWSSecretDBModel getModel(String data) {
-		return GsonUtils.getInstance().stringToModel(data, AWSSecretDBModel.class);
-	}
-}
