+++ sync/aws-secrets-manager-sync/aws-secrets-manager-sync.gradle
@@ -11,19 +11,12 @@ repositories {
     mavenCentral()
 }
 
-
 dependencies {
-
     compile project(":secrets-manager-api")
     compile 'software.amazon.awssdk:secretsmanager'
     compile'com.google.code.gson:gson:2.8.6'
     implementation 'software.amazon.awssdk:sts'
-    testCompile group: 'org.mockito', name: 'mockito-core', version: '3.4.6'
-    testCompile group: 'org.powermock', name: 'powermock-module-junit4', version: '2.0.7'
-    testCompile group: 'org.powermock', name: 'powermock-api-mockito2', version: '2.0.7'
-    testCompile group: 'junit', name: 'junit', version: '4.12'
 }
-
 
 def pomConfig = {
     licenses {
@@ -56,8 +49,6 @@ def pomConfig = {
     }
 }
 
-
-
 publishing {
     publications {
         mavenJava(MavenPublication) {
@@ -75,7 +66,6 @@ publishing {
         }
     }
 }
-
 
 bintray {
     user = project.hasProperty('bintrayUser') ? project.property('bintrayUser') : System.getenv('BINTRAY_USER')
