+++ sync/env-secrets-manager/env-secrets-manager.gradle
@@ -11,13 +11,9 @@ repositories {
     mavenCentral()
 }
 
-
 dependencies {
-
     compile project(":secrets-manager-api")
-    testCompile group: 'junit', name: 'junit', version: '4.12'
 }
-
 
 def pomConfig = {
     licenses {
@@ -78,7 +74,6 @@ publishing {
         }
     }
 }
-
 
 bintray {
     user = project.hasProperty('bintrayUser') ? project.property('bintrayUser') : System.getenv('BINTRAY_USER')
