+++ sync/env-secrets-manager/src/main/java/co/com/bancolombia/secretsmanager/connector/EnvConnector.java
@@ -1,0 +1,29 @@
+package co.com.bancolombia.secretsmanager.connector;
+
+import co.com.bancolombia.secretsmanager.api.GenericManager;
+import co.com.bancolombia.secretsmanager.api.exceptions.SecretException;
+
+/**
+ * Represents an Environment Connector. It lets you to get secrets in
+ * environment variables.
+ *
+ * @author <a href="mailto:andmagom@bancolombia.com.co">Andrés Mauricio Gómez
+ * P.</a>
+ */
+public class EnvConnector implements GenericManager {
+
+    @Override
+    public String getSecret(String secretName) throws SecretException {
+        String myEnv = System.getenv(secretName);
+        if (myEnv == null) {
+            throw new SecretException("The requested secret " + secretName + " was not found");
+        } else {
+            return myEnv;
+        }
+    }
+
+    @Override
+    public <T> T getSecret(String secretName, Class<T> cls){
+        throw new UnsupportedOperationException("Serialization doesn't apply for env connector");
+    }
+}
