+++ sync/file-secrets-manager/file-secrets-manager.gradle
@@ -11,15 +11,9 @@ repositories {
     mavenCentral()
 }
 
-
-
 dependencies {
-
     compile project(":secrets-manager-api")
-    testCompile group: 'junit', name: 'junit', version: '4.12'
 }
-
-
 
 def pomConfig = {
     licenses {
@@ -63,7 +57,6 @@ artifacts {
     archives javadocJar
 }
 
-
 publishing {
     publications {
         mavenJava(MavenPublication) {
@@ -81,7 +74,6 @@ publishing {
         }
     }
 }
-
 
 bintray {
     user = project.hasProperty('bintrayUser') ? project.property('bintrayUser') : System.getenv('BINTRAY_USER')
