+++ sync/file-secrets-manager/src/test/resources/Secret.txt
@@ -1,0 +1,1 @@
+secret
